+++ src/app/controllers/influxTargetCtrl.js
@@ -11,9 +11,8 @@ function (angular) {
   module.controller('InfluxTargetCtrl', function($scope, $timeout) {
 
     $scope.init = function() {
-      if (!$scope.target.function) {
-        $scope.target.function = 'mean';
-      }
+      $scope.target.function = $scope.target.function || 'mean';
+      $scope.target.column = $scope.target.column || 'value';
 
       $scope.rawQuery = false;
 
@@ -42,6 +41,11 @@ function (angular) {
         $scope.oldSeries = $scope.target.series;
         $scope.get_data();
       }
+    };
+
+    $scope.changeFunction = function(func) {
+      $scope.target.function = func;
+      $scope.get_data();
     };
 
     // called outside of digest